REFACTOR: update file structure and imports moved notifications and reviews queries into their own directory and exported each through its index.js moved review controller to its own file and exported it through index refactored all require statements to reflect new file structure #113
