+++ server/database/queries/notification/createNotification.js
@@ -1,0 +1,5 @@
+const Notification = require("../../models/Notification");
+
+const createNotification = data => Notification.create(data);
+
+module.exports = createNotification;
