+++ server/database/queries/notification/index.js
@@ -1,0 +1,6 @@
+const createNotification = require("./createNotification");
+
+const notificationQueries = {};
+module.exports = notificationQueries;
+
+notificationQueries.createNotification = createNotification;
