--- server/database/queries/notifications.js
@@ -1,6 +1,0 @@
-const Notification = require("../models/Notification");
-
-const notifications = {};
-module.exports = notifications;
-
-notifications.createNotification = data => Notification.create(data);
