+++ server/database/queries/review/createReview.js
@@ -1,0 +1,5 @@
+const Review = require("../../models/Review");
+
+const createReview = data => Review.create(data);
+
+module.exports = createReview;
