+++ server/database/queries/review/index.js
@@ -1,0 +1,7 @@
+const createReview = require("./createReview");
+
+const reviewQueries = {};
+module.exports = reviewQueries;
+
+// exports the createReview query
+reviewQueries.createReview = createReview;
