--- server/database/queries/reviews.js
@@ -1,6 +1,0 @@
-const Review = require("../models/Review");
-
-const reviews = {};
-module.exports = reviews;
-
-reviews.createReview = data => Review.create(data);
